workspace/langgraph/studio/requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-openai
langchain-ollama